refactor: move egl-related code in it's own module
